@@ -1,6 +1,6 @@
 {
     "name": "marbie77/emailcountdown",
-    "version": "1.0.0",
+    "version": "2.0.1",
     "type": "library",
     "description": "Email Countdown class to generate GIF",
     "keywords": ["email","countdown"],
@@ -13,9 +13,9 @@
         }
     ],
 	"require": {
-	    "php": ">=5.3.0",
+	    "php": ">=7.1.0",
 	    "ext-gd": "*",
-		"sybio/gif-creator": "1.0.0"
+		"sybio/gif-creator": "^1.0.0"
     },
     "autoload": {
         "psr-4": {
